+++ hahomematic/central_unit.py
@@ -9,13 +9,13 @@ import asyncio
 from collections.abc import Awaitable, Callable, Coroutine
 from concurrent.futures._base import CancelledError
 from datetime import datetime
-import json
 import logging
 import socket
 import threading
 from typing import Any, Final, TypeVar
 
 from aiohttp import ClientSession
+import orjson
 
 from hahomematic import client as hmcl, config
 from hahomematic.caches.dynamic import DeviceDataCache, DeviceDetailsCache
@@ -546,7 +546,7 @@ class CentralUnit:
             await self.paramset_descriptions.load()
             await self.device_details.load()
             await self.device_data.load()
-        except json.decoder.JSONDecodeError:  # pragma: no cover
+        except orjson.JSONDecodeError:  # pragma: no cover
             _LOGGER.warning("LOAD_CACHES failed: Unable to load caches for %s", self._attr_name)
             await self.clear_all_caches()
 
@@ -806,7 +806,7 @@ class CentralUnit:
         return task
 
     def run_coroutine(self, coro: Coroutine) -> Any:
-        """call coroutine from sync."""
+        """Call coroutine from sync."""
         try:
             return asyncio.run_coroutine_threadsafe(coro, self._loop).result()
         except CancelledError:  # pragma: no cover
@@ -837,11 +837,11 @@ class CentralUnit:
         return False
 
     async def fetch_sysvar_data(self, include_internal: bool = True) -> None:
-        """fetch sysvar data for the hub. #CC."""
+        """Fetch sysvar data for the hub. #CC."""
         await self._hub.fetch_sysvar_data(include_internal=include_internal)
 
     async def fetch_program_data(self, include_internal: bool = False) -> None:
-        """fetch program data for the hub. #CC."""
+        """Fetch program data for the hub. #CC."""
         await self._hub.fetch_program_data(include_internal=include_internal)
 
     async def load_and_refresh_entity_data(