+++ changelog.md
@@ -1,3 +1,9 @@
+# Version 2023.2.11 (2023-02-24)
+
+- Update ruff, fix comments
+- Switch to orjson
+- Fix climate: compare set temperature to target temperature
+
 # Version 2023.2.10 (2023-02-23)
 
 - Use sets for central callbacks